+++ doc/examples/with_widgets_overlay/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: 73275f08155aeca9d4719bf685499957091a7a09
+  channel: master
+
+project_type: app
