+++ doc/examples/with_widgets_overlay/README.md
@@ -1,0 +1,3 @@
+# with_widgets_overlay
+
+A Flame game showcasing how to use the WithWidgetsOverlay feature
