+++ doc/examples/with_widgets_overlay/pubspec.yaml
@@ -1,0 +1,17 @@
+name: with_widgets_overlay
+description: A Flame game showcasing how to use the WithWidgetsOverlay feature
+
+version: 1.0.0+1
+
+environment:
+  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+
+dependencies:
+  flutter:
+    sdk: flutter
+  flame:
+      path: ../../../
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
